+++ CHANGELOG.rst
@@ -7,8 +7,9 @@ HEAD
 ----
 * Fix bug where fetching items from a folder that can contain multiple item types (e.g. the Deleted Items folder) would
   only return one item type.
-* ``Folder.bulk_update()`` was changed to expect a map with ``Item`` instance as key and a list of changed fieldnames on
-  each item as value. E.g.:
+* ``Folder.bulk_delete()`` was moved to ``Account.bulk_delete()``
+* ``Folder.bulk_update()`` was moved to ``Account.bulk_update()`` and changed to expect a map with ``Item`` instance as
+key and a list of changed fieldnames on each item as value. E.g.:
 
   .. code-block:: python
 
@@ -22,7 +23,7 @@ HEAD
       for i, item in enumerate(items):
           item.subject = 'Changed subject' % i
           changed_items[item] = 'subject'
-      account.sent.bulk_update(items=item_changes)
+      account.bulk_update(items=item_changes)
 
 
 1.7.0